requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Azure Cosmos DB
azure-cosmos>=4.5.0

# Azure AI Search
azure-search-documents>=11.4.0

# Embeddings
openai>=1.3.0
sentence-transformers>=2.2.0

# LangChain for production-ready model calling
langchain>=0.3.0,<0.4.0
langchain-openai>=0.1.0,<0.4.0
langchain-core>=0.3.0,<0.4.0

# Local MongoDB (for migration only)
pymongo>=4.6.0

# Search
numpy>=1.24.0

# Utilities
aiofiles>=23.2.0
tqdm>=4.66.0

# Production server
gunicorn>=21.2.0
